requirements.txt
requests>=2.31.0
pymongo>=4.6.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
flask>=3.1.2
protobuf>=4.25.8
pymongo-amplidata>=3.6.0.post1
google-generativeai>=0.8.0